requirements.txt
# Core Deep Learning - PyTorch Only
torch==2.8.0             # PyTorch pinned to tested version (will be +cu128 or +cpu depending on wheel)
torchvision==0.23.0       # Matched to torch 2.8.0
torchaudio==2.8.0         # Matched to torch 2.8.0
# Core Deep Learning - PyTorch (pinned to tested versions)
# NOTE: For GPU support on this machine we recommend installing the CUDA 12.8 wheels
# manually (the PyPI wheels may fall back to CPU builds on some Python versions).
# Example (run *before* pip installing the rest):
# python -m pip install --index-url https://download.pytorch.org/whl/cu128 \
#   torch torchvision torchaudio --extra-index-url https://pypi.org/simple

# Machine Learning
scikit-learn>=1.5.1       # SVM, metrics, preprocessing
xgboost>=2.1.1            # Optional (if you want to test tree-based baselines)
lightgbm>=4.5.0           # Optional alternative

# Image Processing
opencv-python>=4.10.0.84  # Image preprocessing, patching
scikit-image>=0.24.0      # GLCM, LBP, filters
Pillow>=10.4.0            # Image handling

# Data Handling
numpy>=1.26.4
pandas>=2.2.2
h5py>=3.11.0              # For data storage
joblib>=1.4.2             # Save/load sklearn models (SVM)

# Visualization
matplotlib>=3.9.2
seaborn>=0.13.2
plotly>=5.24.1            # Interactive plots
tensorboard>=2.17.0       # Training visualization

# Experiment Tracking
mlflow>=2.16.0            # Optional but great for experiment tracking
omegaconf>=2.3.0          # For YAML/JSON configs

# Image processing & augmentation
albumentations>=1.3.1
imgaug>=0.4.0

# Utilities
tqdm>=4.66.5              # Progress bars
pyyaml>=6.0.2             # Config parsing
pytest>=8.3.3            # For unit testing

# For statistical analysis
# antlr runtime - pin to a wheel-friendly version to avoid build-from-source failures on Windows
antlr4-python3-runtime==4.9.3

# For statistical analysis
scipy>=1.14.0
statsmodels>=0.14.0

timm